Rename Logger files
