+++ Libs/libLogger/Channel.cpp
