+++ Libs/libLogger/Channel.hpp
