+++ Libs/libLogger/Event.cpp
