+++ Libs/libLogger/Template.cpp
