+++ Libs/libLogger/Template.hpp
